requirements.txt
numpy
torch
torchvision
streamlit==1.39.0
pillow
